src/test/java/com/hjx/leetcode/problem0200/SolutionTest.java
package com.hjx.leetcode.problem0200;

import org.junit.Assert;
import org.junit.Test;

import static org.junit.Assert.*;

public class SolutionTest {

    @Test
    public void numIslands() {
        Solution solution = new Solution();
        char[][] grid1 = {
                {'1', '1', '1', '1', '0'},
                {'1', '1', '0', '1', '0'},
                {'1', '1', '0', '0', '0'},
                {'0', '0', '0', '0', '0'}
        };

        char[][] grid2 = {
                {'1'}, {1}
        };

        int ans = solution.numIslands(grid2);
        Assert.assertEquals(1, ans);
    }
}